Remove the 2^32 length limit on abstract vectors
Memories are specified as a vec of bytes, but abstract vecs were
restricted to a max length of 2^32. This is obviously incompatible with
the new, longer memories allowed by memory64.

I could have simply raised the abstract limit to 2^64, but I don't yet
understand the purpose of having this limit in the abstract spec anyway.
The binary encoding and various validation and implementation limits
should all enforce appropriate limits as far as I know. So, I just
removed it.